Ash
Version 1.4

+++ src/cparray.cpp
@@ -3,5 +3,5 @@
 CPArray::CPArray():n(0),dim(nullptr),data(nullptr){
 }
 
-CPArray::CPArray(int n, long long *dim, void *data):n(n),dim(dim),data(data){
+CPArray::CPArray(int n, ll* dim, void *data):n(n),dim(dim),data(data){
 }
